Added new Advanced Search form
Added new file: search-tabbed.html with a list of elements representing
tabs. Each of those elements has data attribute (data-type-name) with
value that will be assigned to text field with search terms on click.

E.g. should you click an element with data-type-name="anyWords" that value
will be applied to text input name attribute, changing it to
name="anyWords", which will then be submitted with the rest of the form.

Also added other neccessary attributes to the rest of form fields, so all
the values would submit successfully